copilot fix

+++ esphome/components/dlms_push/dlms_parser.h
@@ -103,8 +103,8 @@ class DlmsParser {
   bool parse_sequence_(uint8_t type, uint8_t depth = 0);
 
   bool capture_generic_value_(AxdrCaptures &c);
-  bool try_match_patterns_(uint8_t elem_idx);
-  bool match_pattern_(uint8_t elem_idx, const AxdrDescriptorPattern &pat, uint8_t &elements_consumed_at_level0);
+  bool try_match_patterns_(uint32_t elem_idx);
+  bool match_pattern_(uint32_t elem_idx, const AxdrDescriptorPattern &pat, uint8_t &elements_consumed_at_level0);
   void emit_object_(const AxdrDescriptorPattern &pat, const AxdrCaptures &c);
 
   float data_as_float_(DlmsDataType value_type, const uint8_t *ptr, uint8_t len);
